Fixed the 1Hz ISR so that only one of update_time or get_time is true.
The ISR was written so that when get_time was true, update_time
was also true. That caused the time to be updated twice and which
made for a glitch once every few minutes where the time would advance
only to be reset a moment later (because code that polls the flags
set by the ISR runs very often).