Removed HV PS PID controller and added a flashing colon.
Should have been two branches. Anyway, the PID controller is gone.
Use a dedicated HV PS that does not need the MCU.

Added a flashing colon and simplified the use of the clock. There
seems to be no need to limit the use of the I2C bus to read from the
DS3231 to once every 10s; it seems to be fine to read every second.